Correções
algumas mudanças de tipo e correções de lógica

+++ Model TAD_Drone/TAD_Drone.c
@@ -1,6 +1,6 @@
 #include "TAD_Drone.h"
 
-Drone inicializar_drone (Drone *drone, float peso){
+Drone inicializar_drone (Drone *drone, int peso){
     drone->peso_max = peso;
     drone->peso_carregado = 0;
     drone->distancia_total = 0;
