+++ Model TAD_Pacote/TAD_Pacote.c
@@ -1,6 +1,6 @@
 #include "TAD_Pacote.h"
 
-void inicializar_pacote(Pacote *pacote, char conteudo[50], char destinatario[50], float peso, float distancia){
+void inicializar_pacote(Pacote *pacote, char conteudo[50], char destinatario[50], int peso, int distancia){
     strcpy(pacote->conteudo, conteudo);
     strcpy(pacote->destinatario, destinatario);
     pacote->peso = peso;
@@ -15,11 +15,11 @@ void set_destinatario(Pacote *pacote, char destinatario[50]){
     strcpy(pacote->destinatario, destinatario);
 }
 
-void set_peso(Pacote *pacote, float peso){
+void set_peso(Pacote *pacote, int peso){
     pacote->peso = peso;
 }
 
-void set_distancia_endereco(Pacote *pacote, float distancia){
+void set_distancia_endereco(Pacote *pacote, int distancia){
     pacote->distancia = distancia;
 }
 
@@ -31,10 +31,10 @@ char get_destinatario(Pacote *pacote){
     return pacote->destinatario;
 }
 
-float get_peso(Pacote *pacote){
+int get_peso(Pacote *pacote){
     return pacote->peso;
 }
 
-float get_distancia_endereco(Pacote *pacote){
+int get_distancia_endereco(Pacote *pacote){
     return pacote->distancia;
 }
