+++ Model TAD_Pacote/TAD_Pacote.h
@@ -10,24 +10,24 @@ typedef struct{
 
     char conteudo[50];
     char destinatario[50];
-    float peso;
-    float distancia;
+    int peso;
+    int distancia;
 
 }Pacote;
 
 //Inicialização do TAD Pacote
-void inicializar_pacote(Pacote *pacote, char conteudo[50], char destinatario[50], float peso, float distancia);
+void inicializar_pacote(Pacote *pacote, char conteudo[50], char destinatario[50], int peso, int distancia);
 
 //Atribuições de valores
 void set_conteudo(Pacote *pacote, char conteudo[50]);
 void set_destinatario(Pacote *pacote, char destinatario[50]);
-void set_peso(Pacote *pacote, float peso);
-void set_distancia_endereco(Pacote *pacote, float distancia);
+void set_peso(Pacote *pacote, int peso);
+void set_distancia_endereco(Pacote *pacote, int distancia);
 
 //Obtenções de valores
 char get_conteudo(Pacote *pacote);
 char get_destinatario(Pacote *pacote);
-float get_peso(Pacote *pacote);
-float get_distancia_endereco(Pacote *pacote);
+int get_peso(Pacote *pacote);
+int get_distancia_endereco(Pacote *pacote);
 
 #endif
